Move react-middleware into separate module
